Revert "밀리세컨드 통일"
This reverts commit e8b3542ff604436c85f9fe8138dff6b6debbb95c.

+++ include/knot/engine.h
@@ -32,12 +32,12 @@ struct TimerTask {
     /**
      * @brief Time remaining until the next event trigger, in milliseconds.
      */
-    float time;
+    double time;
 
     /**
      * @brief The interval between repeated triggers, in milliseconds.
      */
-    float interval;
+    double interval;
 
     /**
      * @brief The unique identifier of the timer task.
@@ -61,9 +61,9 @@ public:
      *
      * The callback is executed after the scene has been rendered.
      *
-     * @param deltaTime The elapsed time since the previous frame, in milliseconds.
+     * @param deltaTime The elapsed time since the previous frame, in seconds.
      */
-    using RenderLoopCallback = std::function<void(float)>;
+    using RenderLoopCallback = std::function<void(double)>;
 
     Engine() = default;
 
@@ -138,7 +138,7 @@ public:
     /**
      * @brief Gets the elapsed time since the previous frame.
      *
-     * @return The frame delta time in milliseconds.
+     * @return The frame delta time in seconds.
      */
     float getDeltaTime() const {
         return deltaTime;
@@ -226,7 +226,7 @@ public:
      *
      * @return The unique ID assigned to the timer task.
      */
-    int after(float interval, uint32_t eventCode, std::any eventData = {});
+    int after(double interval, uint32_t eventCode, std::any eventData = {});
 
     /**
      * @brief Schedules a repeating user event.
@@ -240,7 +240,7 @@ public:
      *
      * @return The unique ID assigned to the timer task.
      */
-    int repeat(float interval, uint32_t eventCode, std::any eventData = {});
+    int repeat(double interval, uint32_t eventCode, std::any eventData = {});
 
     /**
      * @brief Requests the engine to stop running.
@@ -275,8 +275,8 @@ private:
     bool shouldQuit = false;
 
     bool initialized = false;
-    float deltaTime = 0.0f;
-    double lastFrame = 0.0;
+    double deltaTime = 0.0f;
+    double lastFrame = 0.0f;
 
     Scene* scene = nullptr;
